+++ .vscode/launch.json
@@ -11,7 +11,11 @@
             "skipFiles": [
                 "<node_internals>/**"
             ],
-            "program": "${workspaceFolder}/server.js"
+            "cwd": "${workspaceFolder}",
+            "runtimeExecutable": "vercel",
+            "runtimeArgs": [
+                "dev"
+            ],
         }
     ]
 }
