+++ api/languages.js
@@ -1,0 +1,9 @@
+const { languages } = require('../languages');
+
+module.exports = (request, response) => {
+    console.log('');
+    console.log('🎉 ', request.url);
+    response.json({
+        languages
+    });
+}
