+++ languages.js
@@ -1,0 +1,12 @@
+export const languages = [
+    'c', 
+    'css', 
+    'cpp', 
+    'go', 
+    'html', 
+    'java', 
+    'javascript',
+    'python',
+    'rust',
+    'typescript'
+];
