+++ package.json
@@ -10,9 +10,7 @@
   "author": "",
   "license": "GPL V3",
   "dependencies": {
-    "express": "^4.17.1",
-    "nightmare": "^3.0.2",
-    "nightmare-screenshot-selector": "^1.0.2",
-    "puppeteer": "^5.0.0"
+    "puppeteer": "^5.0.0",
+    "vercel": "^19.1.1"
   }
 }
