+++ public/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="refresh" content="0; URL='https://github.com/cyberpirate92/code2img/blob/master/README.md'"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Code2Img</title>
+</head>
+<body>
+    Redirecting to <a href="https://github.com/cyberpirate92/code2img/blob/master/README.md">cyberpirate92/code2img</a>
+</body>
+</html>
